+++ packages/react-components-lab/src/components/CalendarItem/types.ts
@@ -2,7 +2,8 @@ import { ReactNode } from 'react';
 
 export interface CalendarItemProps {
   children: ReactNode;
-  active?: boolean;
   variant: 'button' | 'semi-button';
+  active?: boolean;
   onClick?: () => void;
+  disabled?: boolean;
 }