+++ packages/react-components/src/Scenarios/Scenarios/Scenarios.tsx
@@ -23,23 +23,29 @@ import useStyles from './useStyles';
 const Scenarios = (props: ScenariosProps) => {
   const {
     host,
+    jobHost,
     token,
+    jobToken,
     scenarioConnection,
     queryBody,
-    nameField,
     jobConnection,
     jobParameters,
+    jobQueryItemKey,
     module,
     dataFilterbyProperty,
     signalRConnectionHubUrl,
     taskId,
     hostGroup,
+    nameField,
     descriptionFields,
     extraFields,
     menuItems,
     actionButton,
+    showReportButton,
+    showEditButton,
     selectedScenarioId,
-    showDate,
+    showDate = true,
+    showDateGroups = true,
     showHour,
     showMenu,
     showStatus,
@@ -50,6 +56,9 @@ const Scenarios = (props: ScenariosProps) => {
     onScenarioSelected,
     onScenarioReceived,
     onScenariosReceived,
+    onRenderScenarioItem,
+    onRenderScenarioIcon,
+    onRowRefsUpdated,
     addScenario,
     translations,
     timeZone,
@@ -135,18 +144,20 @@ const Scenarios = (props: ScenariosProps) => {
 
         const query = [
           {
-            item: 'ScenarioId',
+            item: jobQueryItemKey,
             queryOperator: 'Any',
             values,
           },
         ];
-        const dataSources = {
-          host,
+
+        const jobSources = {
+          token: token || jobToken,
+          host: jobHost || host,
           connection: jobConnection,
         };
 
         try {
-          executeJobQuery(dataSources, token, query).subscribe((jobs) => {
+          executeJobQuery(jobSources, query).subscribe((jobs) => {
             newScenarios.map((scenario) => {
               const latestJob = filterToLastJob(scenario, jobs);
               let sce = {};
@@ -183,7 +194,7 @@ const Scenarios = (props: ScenariosProps) => {
   const filterToLastJob = (scenario, jobs) => {
     if (jobs) {
       const latestJobByScenario = jobs
-        .filter((job) => job.data.parameters.ScenarioId === scenario.fullName)
+        .filter((job) => job.data.parameters[jobQueryItemKey] === scenario.fullName)
         .sort((a, b) => b.requested - a.requested);
 
       return latestJobByScenario[0];
@@ -208,9 +219,9 @@ const Scenarios = (props: ScenariosProps) => {
   const onExecuteScenario = (scenario: Scenario, menuItem: MenuItem) => {
     closeDialog();
 
-    // Define Job Parameter with ScenarioId
+    // Define Job Parameter with jobQueryItemKey
     const parameters = {
-      ScenarioId: scenario.fullName,
+      [jobQueryItemKey]: scenario.fullName,
     } as JobParameters;
 
     // Append Job Parameters from Menu Item
@@ -225,10 +236,10 @@ const Scenarios = (props: ScenariosProps) => {
 
     executeJob(
       {
-        host,
+        host: jobHost || host,
         connection: menuItem.connection || jobConnection,
       },
-      token,
+      jobToken || token,
       menuItem.taskId || taskId,
       parameters,
       menuItem.hostGroup || hostGroup,
@@ -238,7 +249,7 @@ const Scenarios = (props: ScenariosProps) => {
       }
 
       const newScenarios = scenarios.map((sce) =>
-        sce.fullName === job.parameters.ScenarioId ? { ...sce, lastJob: job } : { ...sce },
+        sce.fullName === job.parameters[jobQueryItemKey] ? { ...sce, lastJob: job } : { ...sce },
       );
 
       setScenarios(newScenarios);
@@ -254,10 +265,10 @@ const Scenarios = (props: ScenariosProps) => {
 
     cancelJob(
       {
-        host,
+        host: jobHost || host,
         connection: menuItem.connection || jobConnection,
       },
-      token,
+      jobToken || token,
       scenario.lastJob.id,
     );
   };
@@ -484,7 +495,7 @@ const Scenarios = (props: ScenariosProps) => {
     }
 
     const updateScenario = latestScenarios.current.map((scenario) =>
-      scenario.fullName === job.Parameters.ScenarioId && scenario.lastJob.status !== 'Completed'
+      scenario.fullName === job.Parameters[jobQueryItemKey] && scenario.lastJob.status !== 'Completed'
         ? {
             ...scenario,
             lastJob: {
@@ -557,7 +568,11 @@ const Scenarios = (props: ScenariosProps) => {
           selectedScenarioId={selectedScenarioId}
           onScenarioSelected={onScenarioSelectedHandler}
           onContextMenuClick={onContextMenuClickHandler}
+          onRenderScenarioItem={onRenderScenarioItem}
+          onRenderScenarioIcon={onRenderScenarioIcon}
+          onRowRefsUpdated={onRowRefsUpdated}
           showDate={showDate}
+          showDateGroups={showDateGroups}
           showHour={showHour}
           showMenu={showMenu}
           showStatus={showStatus}
@@ -565,6 +580,8 @@ const Scenarios = (props: ScenariosProps) => {
           highlightNameOnStatus={highlightNameOnStatus}
           timeZone={timeZone}
           actionButton={actionButton}
+          showReportButton={showReportButton}
+          showEditButton={showEditButton}
         />
       )}
       {dialog && (