+++ packages/react-components-lab/docs/CHANGELOG.md
@@ -1,3 +1,25 @@
+## @dhi/react-components-lab [1.12.3](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.12.2...@dhi/react-components-lab@1.12.3) (2022-01-19)
+
+
+### Bug Fixes
+
+* **hooks:** Update hooks export ([6e75987](https://github.com/DHI/react-components/commit/6e75987adec7df8800a897eef00ef597c6c2bedd))
+* **stories:** Add stories for the hooks ([04930ea](https://github.com/DHI/react-components/commit/04930eacd0760e2e1588628b1b473078b976170e))
+
+## @dhi/react-components-lab [1.12.2](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.12.1...@dhi/react-components-lab@1.12.2) (2022-01-18)
+
+
+### Bug Fixes
+
+* **CalendarItem:** Add disabled prop ([bec23b4](https://github.com/DHI/react-components/commit/bec23b440cd03fb46f8bdd7d5a32794fdf38b09d))
+
+## @dhi/react-components-lab [1.12.1](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.12.0...@dhi/react-components-lab@1.12.1) (2022-01-18)
+
+
+### Bug Fixes
+
+* **CategoricalBarLegend:** update types ([c0ed13d](https://github.com/DHI/react-components/commit/c0ed13dc3baba2bc5c849e4a535096b20199fb9f))
+
 # @dhi/react-components-lab [1.12.0](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.11.0...@dhi/react-components-lab@1.12.0) (2022-01-05)